Fixed : edit error message for category
