+++ EventsExpress/ClientApp/src/actions/delete-category.js
@@ -4,7 +4,6 @@ import get_categories from './category-list';
 export const SET_CATEGORY_DELETE_PENDING = "SET_CATEGORY_DELETE_PENDING";
 export const SET_CATEGORY_DELETE_SUCCESS = "SET_CATEGORY_DELETE_SUCCESS";
 export const SET_CATEGORY_DELETE_ERROR = "SET_CATEGORY_DELETE_ERROR";
-export const SET_CATEGORY_EDITED = "SET_CATEGORY_EDITED";
 
 const api_serv = new CategoryService();
 
@@ -26,21 +25,6 @@ export function delete_category(data) {
     }
 }
 
-export function set_edited_category(id) {
-    return dispatch => {
-        dispatch(setCategoryEdited(id));
-        dispatch(setCategoryDeletePending(false));
-        dispatch(setCategoryDeleteError(null));
-        dispatch(setCategoryDeleteSuccess(false));
-    }
-}
-
-function setCategoryEdited(data) {
-    return {
-        type: SET_CATEGORY_EDITED,
-        payload: data
-    };
-}
 
 function setCategoryDeleteSuccess(data) {
     return {
