+++ EventsExpress/ClientApp/src/reducers/category-list.js
@@ -3,7 +3,7 @@ import initialState from '../store/initialState';
 import {
     SET_CATEGORIES_ERROR, SET_CATEGORIES_PENDING, GET_CATEGORIES_SUCCESS
 } from '../actions/category-list';
-import { SET_CATEGORY_EDITED } from '../actions/delete-category'
+import { SET_CATEGORY_EDITED } from '../actions/add-category'
 
 export const reducer = (state = initialState.categories, action) => {
     switch (action.type) {
